+++ packages/client/ui/react-ui/src/index.ts
@@ -2,6 +2,8 @@ export * from "./components";
 export * from "./hooks";
 export * from "./providers";
 
+export type { LoginMethod } from "./types/auth";
+
 export { CrossmintEvents, useCrossmintEvents } from "@crossmint/client-sdk-base";
 
 export type { EVMSmartWallet, SolanaSmartWallet } from "@crossmint/wallets-sdk";