+++ packages/client/ui/react-ui/src/providers/auth/AuthFormProvider.tsx
@@ -1,9 +1,8 @@
 import { createContext, useCallback, useContext, useEffect, useState, type ReactNode } from "react";
 import { CrossmintAuthenticationError, type OAuthProvider } from "@crossmint/common-sdk-auth";
 import type { UIConfig } from "@crossmint/common-sdk-base";
-import type { CrossmintAuthEmbeddedWallets } from "@/types/wallet";
 import { useCrossmintAuth } from "@/hooks/useCrossmintAuth";
-import type { LoginMethod } from "@/types/auth";
+import type { CrossmintAuthProviderEmbeddedWallets, LoginMethod } from "@/types/auth";
 
 type AuthStep = "initial" | "otp" | "qrCode" | "web3" | "web3/metamask" | "web3/walletconnect";
 
@@ -37,7 +36,7 @@ type ContextInitialStateProps = {
     defaultEmail?: string;
     baseUrl: string;
     setDialogOpen?: (open: boolean, successfulLogin?: boolean) => void;
-    embeddedWallets: CrossmintAuthEmbeddedWallets;
+    embeddedWallets: CrossmintAuthProviderEmbeddedWallets;
 };
 
 type AuthFormProviderProps = {
@@ -69,12 +68,7 @@ export const AuthFormProvider = ({
     const [oauthUrlMap, setOauthUrlMap] = useState<OAuthUrlMap>(initialOAuthUrlMap);
     const [isLoadingOauthUrlMap, setIsLoadingOauthUrlMap] = useState(true);
 
-    const { loginMethods, baseUrl, appearance, embeddedWallets, termsOfServiceText, authModalTitle, defaultEmail } =
-        initialState;
-
-    if (loginMethods.includes("web3") && embeddedWallets?.createOnLogin === "all-users") {
-        throw new Error("Creating wallets on login is not yet supported for web3 login method");
-    }
+    const { loginMethods, baseUrl, appearance, termsOfServiceText, authModalTitle, defaultEmail } = initialState;
 
     const preFetchAndSetOauthUrl = useCallback(async () => {
         setIsLoadingOauthUrlMap(true);