+++ packages/client/ui/react-ui/src/providers/auth/web3/DynamicWeb3WalletConnect.tsx
@@ -1,7 +1,8 @@
 import type React from "react";
-import type { APIKeyEnvironmentPrefix } from "@crossmint/common-sdk-base";
+import base58 from "bs58";
 import { EthereumWalletConnectors } from "@dynamic-labs/ethereum";
 import { SolanaWalletConnectors } from "@dynamic-labs/solana";
+import type { APIKeyEnvironmentPrefix } from "@crossmint/common-sdk-base";
 import DynamicContextProviderWrapper from "@/components/dynamic-xyz/DynamicContextProviderWrapper";
 import { useCrossmintAuth } from "@/hooks/useCrossmintAuth";
 import { useAuthForm } from "../AuthFormProvider";
@@ -10,19 +11,31 @@ import { dynamicChainToCrossmintChain } from "@/utils/dynamic/dynamicChainToCros
 export function DynamicWeb3WalletConnect({
     children,
     apiKeyEnvironment,
-}: { children: React.ReactNode; apiKeyEnvironment: APIKeyEnvironmentPrefix }) {
+    enabled,
+}: { children: React.ReactNode; apiKeyEnvironment: APIKeyEnvironmentPrefix; enabled: boolean }) {
     const { crossmintAuth } = useCrossmintAuth();
-    const { appearance } = useAuthForm();
+    const { appearance, loginMethods } = useAuthForm();
+
+    let connectors = enabled ? [EthereumWalletConnectors, SolanaWalletConnectors] : undefined;
+    if (loginMethods.includes("web3:evm-only")) {
+        connectors = [EthereumWalletConnectors];
+    } else if (loginMethods.includes("web3:solana-only")) {
+        connectors = [SolanaWalletConnectors];
+    }
 
     const cssOverrides = `.powered-by-dynamic { display: none !important; } .wallet-list__scroll-container { padding: 0px !important; } .wallet-list__search-container { padding-left: 0px !important; padding-right: 0px !important; } .dynamic-footer { display: none !important; } h1 { color: ${appearance?.colors?.textPrimary} !important; } * { color: ${appearance?.colors?.textSecondary} !important; }`;
 
     return (
         <DynamicContextProviderWrapper
+            key={`${apiKeyEnvironment}-${loginMethods.join(",")}`}
             apiKeyEnvironment={apiKeyEnvironment}
             settings={{
-                walletConnectors: [EthereumWalletConnectors, SolanaWalletConnectors],
+                walletConnectors: connectors,
                 cssOverrides,
                 events: {
+                    onWalletRemoved() {
+                        console.log("[CryptoWalletConnectionHandler] onWalletRemoved");
+                    },
                     onAuthFlowCancel() {
                         console.log("[CryptoWalletConnectionHandler] onAuthFlowCancel");
                     },
@@ -37,31 +50,40 @@ export function DynamicWeb3WalletConnect({
                     },
                 },
                 handlers: {
-                    handleConnectedWallet: async (wallet) => {
-                        console.log("[CryptoWalletConnectionHandler] handleConnectedWallet", wallet);
-
-                        const address = wallet.address;
+                    handleConnectedWallet: async (args) => {
+                        console.log("[CryptoWalletConnectionHandler] handleConnectedWallet", args);
+                        const address = args.address;
                         if (!address) {
                             console.error("[CryptoWalletConnectionHandler] handleConnectedWallet: address is missing");
                             return false;
                         }
-
-                        const chain = await dynamicChainToCrossmintChain(wallet);
-                        console.log({ chain });
-
+                        const chain = await dynamicChainToCrossmintChain(args);
+                        const type = chain === "solana" ? "solana" : "evm";
                         try {
-                            const res = await crossmintAuth?.signInWithSmartWallet(address);
-                            const signature = (await wallet.signMessage?.(res.challenge)) as `0x${string}`;
-                            const authResponse = (await crossmintAuth?.authenticateSmartWallet(address, signature)) as {
+                            const res = await crossmintAuth?.signInWithSmartWallet(address, type);
+                            const rawSignature = (await args.connector?.proveOwnership(
+                                address,
+                                res.challenge
+                            )) as string;
+
+                            let signature = rawSignature;
+                            if (type === "solana") {
+                                const signatureBuffer = Buffer.from(rawSignature, "base64");
+                                signature = base58.encode(signatureBuffer);
+                            }
+                            const authResponse = (await crossmintAuth?.authenticateSmartWallet(
+                                address,
+                                type,
+                                signature
+                            )) as {
                                 oneTimeSecret: string;
                             };
                             const oneTimeSecret = authResponse.oneTimeSecret;
                             await crossmintAuth?.handleRefreshAuthMaterial(oneTimeSecret);
                         } catch (error) {
                             console.error("[CryptoWalletConnectionHandler] Error authenticating wallet:", error);
-                            return false;
+                            false;
                         }
-
                         return true;
                     },
                 },