+++ apps/wallets/smart-wallet/expo/CHANGELOG.md
@@ -1,5 +1,11 @@
 # expo-demo
 
+## 1.0.2
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-native-ui@0.1.1
+
 ## 1.0.1
 
 ### Patch Changes