+++ packages/client/react-base/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @crossmint/client-sdk-react-base
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [30841fa]
+  - @crossmint/wallets-sdk@0.4.4
+
 ## 0.1.0
 
 ### Minor Changes